// src/app/shared/model/VerfResultHObj.Model.ts
export class VerfResultHObj {
    VerfResultHId: any;
    VerfResultId: any;
    VerfSchemeHId: any;
    MrVerfObjectCode: any;
    MrVerfSubjectRelationCode: any;
    VerfDt: any;
    MrVerfResultHStatCode: any;
    Phn: any;
    PhnType: any;
    Notes: any;
    Version: any;
    RowVersion : any;
    constructor() { this.VerfResultHId = 0; this.VerfResultId = 0; this.VerfSchemeHId = 0; }
}  