// src/app/shared/model/VerfResultObj.Model.ts
export class VerfResultObj {
    VerfResultId: any;
    VerfResultNo: any;
    TrxRefNo: any;
    VerfDt: any;
    EmpNo: any;
    MrVerfResultStatCode: any;
    MrVerfTrxTypeCode: any;
    LobCode: any;
    LobName: any;
    Notes: any;
    RowVersion : any;
    constructor() { this.VerfResultId = 0; }
}  